requirements.txt
numpy==1.18.5
pyzbar==0.1.8
Pillow==8.2.0
PyMuPDF==1.18.13
